+++ sdk/communication/Azure.Communication.Sms/src/Models/SmsRecipient.cs
@@ -6,7 +6,6 @@ using Azure.Core;
 namespace Azure.Communication.Sms.Models
 {
     /// <summary> Sms Recipient. </summary>
-
     [CodeGenModel("SmsRecipient")]
     internal partial class SmsRecipient
     {