+++ sdk/communication/Azure.Communication.Sms/src/Models/SmsSendResponse.cs
@@ -5,6 +5,7 @@ using Azure.Core;
 
 namespace Azure.Communication.Sms.Models
 {
+    /// <summary> SMS Send Response. </summary>
     [CodeGenModel("SmsSendResponse")]
     internal partial class SmsSendResponse
     {