+++ sdk/communication/Azure.Communication.Sms/src/Models/SmsRecipient.cs
@@ -5,8 +5,8 @@ using Azure.Core;
 
 namespace Azure.Communication.Sms.Models
 {
-    [CodeGenModel("SendMessageRequest")]
-    internal partial class SendMessageRequest
+    [CodeGenModel("SmsRecipient")]
+    internal partial class SmsRecipient
     {
     }
 }