+++ sdk/communication/Azure.Communication.Sms/tests/SmsClientRecordedTestSanitizer.cs
@@ -11,12 +11,14 @@ namespace Azure.Communication.Sms.Tests
         {
             JsonPathSanitizers.Add("$..from");
             JsonPathSanitizers.Add("$..to");
+            JsonPathSanitizers.Add("$..repeatabilityRequestId");
+            JsonPathSanitizers.Add("$..repeatabilityFirstSent");
         }
 
         public override string SanitizeVariable(string variableName, string environmentVariableValue)
             => variableName switch
             {
-                SmsClientTestEnvironment.PhoneNumberEnvironmentVariableName => "+18005555555",
+                SmsClientTestEnvironment.FromPhoneNumberEnvironmentVariableName => "+18005555555",
                 _ => base.SanitizeVariable(variableName, environmentVariableValue)
             };
     }