+++ sdk/communication/Azure.Communication.Sms/CHANGELOG.md
@@ -11,7 +11,7 @@
 
 ### Breaking
 - Updated `Task<Response<SendSmsResponse>> SendAsync(PhoneNumberIdentifier from, PhoneNumberIdentifier to, string message, SendSmsOptions sendSmsOptions = null, CancellationToken cancellationToken = default)`
-to `Task<Response<SmsSendResult>> SendAsync(string from, string to, string message, Models.SmsSendOptions options = default)`
+to `Task<Response<SmsSendResult>> SendAsync(string from, string to, string message, SmsSendOptions options = default)`
 - Replaced `SendSmsResponse` with `SmsSendResult`
 
 ## 1.0.0-beta.3 (2020-11-16)