+++ sdk/communication/Azure.Communication.Sms/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 1.0.0-beta.4 (2021-03-09)
+## 1.0.0-beta.5 (2021-03-09)
 
 ### Added
 - Added support to create SmsClient with AzureKeyCredential.